update type amount to float

+++ ManagementSystem.Common/Models/Dtos/EmployeeShifting/InventoryAuditDetailDto.cs
@@ -12,7 +12,7 @@ namespace ManagementSystem.Common.Models.Dtos
         public string? ProductName { get; set; }
         public int? UnitId { get; set; }
         public string? UnitName { get; set; }
-        public int? ActualAmount { get; set; }
-        public int? SystemAmount { get; set; }
+        public float? ActualAmount { get; set; }
+        public float? SystemAmount { get; set; }
     }
 }
